ftpm: Add QEMU v7 fTPM implementation
This hangs when fTPM TA is about to use secure storage. This is because
we are running this before Linux kernel has booted up. I.e.,
tee-supplicant and other vital parts are not running. We should probably
try to start run the fTPM driver as a module after we have successfully
booted up Linux.

To test this:
1.   cd build && make optee-os CFG_EARLY_TA=y ta_dev_kit

2.   cd ms-tpm-20-ref
2.1  git submodule init
2.2  git submodule update
2.3  make

3.   cd build && make -j8 CFG_EARLY_TA=y EARLY_TA_PATHS=/media/jbech/SSHD_LINUX/devel/optee_projects/qemu/ms-tpm-20-ref/out/bc50d971-d4c9-42c4-82cb-123456789123.stripped.elf run

To load the TA in GDB:
1. Run once to see the load address (D/TC:? 0 load_elf_from_store:801 ELF load address 0x113000)
2. add-symbol-file /media/jbech/SSHD_LINUX/devel/optee_projects/qemu/ms-tpm-20-ref/out/bc50d971-d4c9-42c4-82cb-123456789123.elf 0x113020
3. b TA_CreateEntryPoint

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>